requirements.txt
rich
spotipy
yt-dlp
pydub
python-dotenv
requests 